+++ packages/common/src/httpRequests.ts
@@ -1,3 +1,4 @@
+import fetch from "node-fetch";
 import { Logger } from "./Logger";
 
 export enum HttpMethod {