+++ packages/web3-auth-native/src/index.ts
@@ -1,6 +1,0 @@
-import SocialLogin from './SocialLogin'
-
-export default SocialLogin
-export * from './types/IWebBrowser'
-export * from './types/sdk'
-export * from './types/State'