+++ packages/web3-auth/src/index.ts
@@ -1,4 +1,0 @@
-import SocialLogin, { socialLoginSDK, getSocialLoginSDK } from "./SocialLogin";
-
-export default SocialLogin;
-export { socialLoginSDK, getSocialLoginSDK };