+++ packages/bundler/src/Bundler.ts
@@ -1,6 +1,6 @@
 import { IBundler } from "./interfaces/IBundler";
 import {
-  GetUserOperationResponse,
+  GetUserOperationReceiptResponse,
   GetUserOpByHashResponse,
   Bundlerconfig,
   UserOpResponse,
@@ -9,14 +9,21 @@ import {
   SendUserOpResponse,
   UserOpGasResponse,
   UserOpByHashResponse,
-  SendUserOpOptions,
   GetGasFeeValuesResponse,
   GasFeeValues,
+  UserOpStatus,
+  GetUserOperationStatusResponse,
+  SimulationType,
 } from "./utils/Types";
 import { resolveProperties } from "ethers/lib/utils";
 import { deepHexlify, sendRequest, getTimestampInSeconds, HttpMethod, Logger, RPC_PROVIDER_URLS } from "@biconomy/common";
 import { transformUserOP } from "./utils/HelperFunction";
-import { UserOpReceiptIntervals } from "./utils/Constants";
+import {
+  UserOpReceiptIntervals,
+  UserOpWaitForTxHashIntervals,
+  UserOpWaitForTxHashMaxDurationIntervals,
+  UserOpReceiptMaxDurationIntervals,
+} from "./utils/Constants";
 import { JsonRpcProvider } from "@ethersproject/providers";
 import { type UserOperationStruct } from "@alchemy/aa-core";
 
@@ -27,12 +34,33 @@ import { type UserOperationStruct } from "@alchemy/aa-core";
  */
 export class Bundler implements IBundler {
   // eslint-disable-next-line no-unused-vars
-  UserOpReceiptIntervals: { [key in number]?: number };
+  UserOpReceiptIntervals!: { [key in number]?: number };
+
+  UserOpWaitForTxHashIntervals!: { [key in number]?: number };
+
+  UserOpReceiptMaxDurationIntervals!: { [key in number]?: number };
+
+  UserOpWaitForTxHashMaxDurationIntervals!: { [key in number]?: number };
 
   constructor(readonly bundlerConfig: Bundlerconfig) {
     this.UserOpReceiptIntervals = {
       ...UserOpReceiptIntervals,
       ...bundlerConfig.userOpReceiptIntervals,
+    };
+
+    this.UserOpWaitForTxHashIntervals = {
+      ...UserOpWaitForTxHashIntervals,
+      ...bundlerConfig.userOpWaitForTxHashIntervals,
+    };
+
+    this.UserOpReceiptMaxDurationIntervals = {
+      ...UserOpReceiptMaxDurationIntervals,
+      ...bundlerConfig.userOpReceiptMaxDurationIntervals,
+    };
+
+    this.UserOpWaitForTxHashMaxDurationIntervals = {
+      ...UserOpWaitForTxHashMaxDurationIntervals,
+      ...bundlerConfig.userOpWaitForTxHashMaxDurationIntervals,
     };
   }
 
@@ -81,13 +109,13 @@ export class Bundler implements IBundler {
    * @description This function will send signed userOp to bundler to get mined on chain
    * @returns Promise<UserOpResponse>
    */
-  async sendUserOp(userOp: UserOperationStruct, simulationParam?: SendUserOpOptions): Promise<UserOpResponse> {
+  async sendUserOp(userOp: UserOperationStruct, simulationParam?: SimulationType): Promise<UserOpResponse> {
     const chainId = this.bundlerConfig.chainId;
     // transformUserOP will convert all bigNumber values to string
     userOp = transformUserOP(userOp);
     const hexifiedUserOp = deepHexlify(await resolveProperties(userOp));
     const simType = {
-      simulation_type: simulationParam?.simulationType || "validation",
+      simulation_type: simulationParam || "validation",
     };
     const params = [hexifiedUserOp, this.bundlerConfig.entryPointAddress, simType];
     const bundlerUrl = this.getBundlerUrl();
@@ -105,7 +133,12 @@ export class Bundler implements IBundler {
       userOpHash: sendUserOperationResponse.result,
       wait: (confirmations?: number): Promise<UserOpReceipt> => {
         const provider = new JsonRpcProvider(RPC_PROVIDER_URLS[chainId]);
+        // Note: maxDuration can be defined per chainId
+        const maxDuration = this.UserOpReceiptMaxDurationIntervals[chainId] || 30000; // default 30 seconds
+        let totalDuration = 0;
+
         return new Promise<UserOpReceipt>((resolve, reject) => {
+          const intervalValue = this.UserOpReceiptIntervals[chainId] || 5000; // default 5 seconds
           const intervalId = setInterval(async () => {
             try {
               const userOpResponse = await this.getUserOpReceipt(sendUserOperationResponse.result);
@@ -125,7 +158,52 @@ export class Bundler implements IBundler {
               clearInterval(intervalId);
               reject(error);
             }
-          }, this.UserOpReceiptIntervals[chainId]);
+
+            totalDuration += intervalValue;
+            if (totalDuration >= maxDuration) {
+              clearInterval(intervalId);
+              reject(
+                new Error(
+                  `Exceeded maximum duration (${maxDuration / 1000} sec) waiting to get receipt for userOpHash ${
+                    sendUserOperationResponse.result
+                  }. Try getting the receipt manually using eth_getUserOperationReceipt rpc method on bundler`,
+                ),
+              );
+            }
+          }, intervalValue);
+        });
+      },
+      waitForTxHash: (): Promise<UserOpStatus> => {
+        const maxDuration = this.UserOpWaitForTxHashMaxDurationIntervals[chainId] || 20000; // default 20 seconds
+        let totalDuration = 0;
+
+        return new Promise<UserOpStatus>((resolve, reject) => {
+          const intervalValue = this.UserOpWaitForTxHashIntervals[chainId] || 500; // default 0.5 seconds
+          const intervalId = setInterval(() => {
+            this.getUserOpStatus(sendUserOperationResponse.result)
+              .then((userOpStatus) => {
+                if (userOpStatus && userOpStatus.state && userOpStatus.transactionHash) {
+                  clearInterval(intervalId);
+                  resolve(userOpStatus);
+                }
+              })
+              .catch((error) => {
+                clearInterval(intervalId);
+                reject(error);
+              });
+
+            totalDuration += intervalValue;
+            if (totalDuration >= maxDuration) {
+              clearInterval(intervalId);
+              reject(
+                new Error(
+                  `Exceeded maximum duration (${maxDuration / 1000} sec) waiting to get receipt for userOpHash ${
+                    sendUserOperationResponse.result
+                  }. Try getting the receipt manually using eth_getUserOperationReceipt rpc method on bundler`,
+                ),
+              );
+            }
+          }, intervalValue);
         });
       },
     };
@@ -140,7 +218,7 @@ export class Bundler implements IBundler {
    */
   async getUserOpReceipt(userOpHash: string): Promise<UserOpReceipt> {
     const bundlerUrl = this.getBundlerUrl();
-    const response: GetUserOperationResponse = await sendRequest({
+    const response: GetUserOperationReceiptResponse = await sendRequest({
       url: bundlerUrl,
       method: HttpMethod.Post,
       body: {
@@ -152,6 +230,28 @@ export class Bundler implements IBundler {
     });
     const userOpReceipt: UserOpReceipt = response.result;
     return userOpReceipt;
+  }
+
+  /**
+   *
+   * @param userOpHash
+   * @description This function will return userOpReceipt for a given userOpHash
+   * @returns Promise<UserOpReceipt>
+   */
+  async getUserOpStatus(userOpHash: string): Promise<UserOpStatus> {
+    const bundlerUrl = this.getBundlerUrl();
+    const response: GetUserOperationStatusResponse = await sendRequest({
+      url: bundlerUrl,
+      method: HttpMethod.Post,
+      body: {
+        method: "biconomy_getUserOperationStatus",
+        params: [userOpHash],
+        id: getTimestampInSeconds(),
+        jsonrpc: "2.0",
+      },
+    });
+    const userOpStatus: UserOpStatus = response.result;
+    return userOpStatus;
   }
 
   /**