+++ packages/bundler/src/interfaces/IBundler.ts
@@ -1,10 +1,11 @@
-import { UserOpResponse, UserOpGasResponse, UserOpReceipt, UserOpByHashResponse, SendUserOpOptions, GasFeeValues } from "../utils/Types";
+import { UserOpResponse, UserOpGasResponse, UserOpReceipt, UserOpByHashResponse, UserOpStatus, SimulationType, GasFeeValues } from "../utils/Types";
 import { UserOperationStruct } from "@alchemy/aa-core";
 
 export interface IBundler {
   estimateUserOpGas(_userOp: Partial<UserOperationStruct>): Promise<UserOpGasResponse>;
-  sendUserOp(_userOp: UserOperationStruct, _simulationParam?: SendUserOpOptions): Promise<UserOpResponse>;
+  sendUserOp(_userOp: UserOperationStruct, _simulationType?: SimulationType): Promise<UserOpResponse>;
   getUserOpReceipt(_userOpHash: string): Promise<UserOpReceipt>;
   getUserOpByHash(_userOpHash: string): Promise<UserOpByHashResponse>;
   getGasFeeValues(): Promise<GasFeeValues>;
+  getUserOpStatus(_userOpHash: string): Promise<UserOpStatus>;
 }