+++ packages/bundler/src/utils/Types.ts
@@ -7,6 +7,9 @@ export type Bundlerconfig = {
   chainId: ChainId;
   // eslint-disable-next-line no-unused-vars
   userOpReceiptIntervals?: { [key in ChainId]?: number };
+  userOpWaitForTxHashIntervals?: { [key in ChainId]?: number };
+  userOpReceiptMaxDurationIntervals?: { [key in ChainId]?: number };
+  userOpWaitForTxHashMaxDurationIntervals?: { [key in ChainId]?: number };
 };
 
 export type UserOpReceipt = {
@@ -23,17 +26,27 @@ export type UserOpReceipt = {
   receipt: ethers.providers.TransactionReceipt;
 };
 
-export type SendUserOpOptions = {
-  simulationType?: SimulationType;
+// review
+export type UserOpStatus = {
+  state: string; // for now // could be an enum
+  transactionHash?: string;
+  userOperationReceipt?: UserOpReceipt;
 };
 
 export type SimulationType = "validation" | "validation_and_execution";
 
 // Converted to JsonRpcResponse with strict type
-export type GetUserOperationResponse = {
+export type GetUserOperationReceiptResponse = {
   jsonrpc: string;
   id: number;
   result: UserOpReceipt;
+  error?: JsonRpcError;
+};
+
+export type GetUserOperationStatusResponse = {
+  jsonrpc: string;
+  id: number;
+  result: UserOpStatus;
   error?: JsonRpcError;
 };
 
@@ -48,6 +61,8 @@ export type SendUserOpResponse = {
 export type UserOpResponse = {
   userOpHash: string;
   wait(_confirmations?: number): Promise<UserOpReceipt>;
+  // Review: waitForTxHash(): vs waitForTxHash?():
+  waitForTxHash(): Promise<UserOpStatus>;
 };
 
 // Converted to JsonRpcResponse with strict type