+++ packages/core-types/src/ChainsTypes.ts
@@ -1,31 +1,3 @@
-export enum ChainNames {
-  Mainnet = "mainnet",
-  Ropsten = "ropsten",
-  Rinkeby = "rinkeby",
-  Goerli = "goerli",
-  Kovan = "kovan",
-  Xdai = "xdai",
-  Bsc = "bsc",
-  BscTest = "bscTest",
-  Fantom = "fantom",
-  FantomTest = "fantomTest",
-  Matic = "matic",
-  Mumbai = "mumbai",
-  Aurora = "aurora",
-  AuroraTest = "auroraTest",
-  Avalanche = "avalanche",
-  Fuji = "fuji",
-  Optimism = "optimism",
-  OptimismKovan = "optimismKovan",
-  Arbitrum = "arbitrum",
-  ArbitrumTest = "arbitrumTest",
-  Moonbeam = "moonbeam",
-  Moonbase = "moonbase",
-  Celo = "celo",
-  CeloTest = "celoTest",
-}
-
-// NOTE: Update chainId for networks we're planning to support
 export enum ChainId {
   // Ethereum
   MAINNET = 1,
@@ -47,9 +19,14 @@ export enum ChainId {
   BASE_GOERLI_TESTNET = 84531,
   BASE_MAINNET = 8453,
   LINEA_TESTNET = 59140,
+  LINEA_MAINNET = 59144,
   MANTLE_MAINNET = 5000,
   MANTLE_TESTNET = 5001,
   OPBNB_MAINNET = 204,
   OPBNB_TESTNET = 5611,
+  ASTAR_MAINNET = 592,
+  ASTAR_TESTNET = 81,
+  CHILIZ_MAINNET = 88888,
+  CHILIZ_TESTNET = 88882,
   GANACHE = 1337, //Temp
 }